Remove `ActionMap`
Use Bevy's `TypeIdMap` instead.
And merge `acton_map` module into `input_action`.
