+++ src/input_condition/block_by.rs
@@ -1,8 +1,12 @@
-use core::{any, marker::PhantomData};
+use core::{
+    any::{self, TypeId},
+    marker::PhantomData,
+};
 
+use bevy::utils::TypeIdMap;
 use log::warn;
 
-use crate::{action_map::ActionMap, prelude::*};
+use crate::prelude::*;
 
 /// Requires another action to not be fired within the same context.
 #[derive(Debug)]
@@ -30,11 +34,11 @@ impl<A: InputAction> Copy for BlockBy<A> {}
 impl<A: InputAction> InputCondition for BlockBy<A> {
     fn evaluate(
         &mut self,
-        action_map: &ActionMap,
+        action_map: &TypeIdMap<Action>,
         _time: &InputTime,
         _value: ActionValue,
     ) -> ActionState {
-        if let Some(action) = action_map.action::<A>() {
+        if let Some(action) = action_map.get(&TypeId::of::<A>()) {
             if action.state() == ActionState::Fired {
                 return ActionState::None;
             }
@@ -56,8 +60,6 @@ impl<A: InputAction> InputCondition for BlockBy<A> {
 
 #[cfg(test)]
 mod tests {
-    use core::any::TypeId;
-
     use bevy_enhanced_input_macros::InputAction;
 
     use super::*;
@@ -71,7 +73,7 @@ mod tests {
         let time = state.get(&world);
 
         action.update(&time, ActionState::Fired, true);
-        let mut action_map = ActionMap::default();
+        let mut action_map = TypeIdMap::<Action>::default();
         action_map.insert(TypeId::of::<TestAction>(), action);
 
         assert_eq!(
@@ -83,7 +85,7 @@ mod tests {
     #[test]
     fn missing_action() {
         let mut condition = BlockBy::<TestAction>::default();
-        let action_map = ActionMap::default();
+        let action_map = TypeIdMap::<Action>::default();
         let (world, mut state) = input_time::init_world();
         let time = state.get(&world);
