+++ src/input_modifier.rs
@@ -9,9 +9,10 @@ pub mod smooth_nudge;
 pub mod swizzle_axis;
 
 use alloc::boxed::Box;
+use bevy::utils::TypeIdMap;
 use core::{fmt::Debug, iter};
 
-use crate::{action_map::ActionMap, prelude::*};
+use crate::prelude::*;
 
 /// Pre-processor that alter the raw input values.
 ///
@@ -26,7 +27,7 @@ pub trait InputModifier: Sync + Send + Debug + 'static {
     /// Called each frame.
     fn apply(
         &mut self,
-        action_map: &ActionMap,
+        action_map: &TypeIdMap<Action>,
         time: &InputTime,
         value: ActionValue,
     ) -> ActionValue;
