+++ src/input_modifier/accumulate_by.rs
@@ -1,9 +1,12 @@
-use core::{any, marker::PhantomData};
+use core::{
+    any::{self, TypeId},
+    marker::PhantomData,
+};
 
-use bevy::prelude::*;
+use bevy::{prelude::*, utils::TypeIdMap};
 use log::warn;
 
-use crate::{action_map::ActionMap, prelude::*};
+use crate::prelude::*;
 
 /// Produces accumulated value when another action is fired within the same context.
 ///
@@ -30,11 +33,11 @@ impl<A: InputAction> Default for AccumulateBy<A> {
 impl<A: InputAction> InputModifier for AccumulateBy<A> {
     fn apply(
         &mut self,
-        action_map: &ActionMap,
+        action_map: &TypeIdMap<Action>,
         _time: &InputTime,
         value: ActionValue,
     ) -> ActionValue {
-        if let Some(action) = action_map.action::<A>() {
+        if let Some(action) = action_map.get(&TypeId::of::<A>()) {
             if action.state() == ActionState::Fired {
                 self.value += value.as_axis3d();
             } else {
@@ -54,8 +57,6 @@ impl<A: InputAction> InputModifier for AccumulateBy<A> {
 
 #[cfg(test)]
 mod tests {
-    use core::any::TypeId;
-
     use bevy_enhanced_input_macros::InputAction;
 
     use super::*;
@@ -69,7 +70,7 @@ mod tests {
         let time = state.get(&world);
 
         action.update(&time, ActionState::Fired, true);
-        let mut action_map = ActionMap::default();
+        let mut action_map = TypeIdMap::<Action>::default();
         action_map.insert(TypeId::of::<TestAction>(), action);
 
         assert_eq!(modifier.apply(&action_map, &time, 1.0.into()), 1.0.into());
@@ -82,7 +83,7 @@ mod tests {
         let action = Action::new::<TestAction>();
         let (world, mut state) = input_time::init_world();
         let time = state.get(&world);
-        let mut action_map = ActionMap::default();
+        let mut action_map = TypeIdMap::<Action>::default();
         action_map.insert(TypeId::of::<TestAction>(), action);
 
         assert_eq!(modifier.apply(&action_map, &time, 1.0.into()), 1.0.into());
@@ -92,7 +93,7 @@ mod tests {
     #[test]
     fn missing_action() {
         let mut modifier = AccumulateBy::<TestAction>::default();
-        let action_map = ActionMap::default();
+        let action_map = TypeIdMap::<Action>::default();
         let (world, mut state) = input_time::init_world();
         let time = state.get(&world);
