+++ src/input_modifier/delta_scale.rs
@@ -1,4 +1,6 @@
-use crate::{action_map::ActionMap, prelude::*};
+use bevy::utils::TypeIdMap;
+
+use crate::prelude::*;
 
 /// Multiplies the input value by delta time for this frame.
 ///
@@ -9,7 +11,7 @@ pub struct DeltaScale;
 impl InputModifier for DeltaScale {
     fn apply(
         &mut self,
-        _action_map: &ActionMap,
+        _action_map: &TypeIdMap<Action>,
         time: &InputTime,
         value: ActionValue,
     ) -> ActionValue {
@@ -36,7 +38,7 @@ mod tests {
 
     #[test]
     fn scaling() {
-        let action_map = ActionMap::default();
+        let action_map = TypeIdMap::<Action>::default();
         let (mut world, mut state) = input_time::init_world();
         world
             .resource_mut::<Time>()
